m2x v2.0.0 M2X.Collection

Wrapper for the AT&T M2X Collection API

Functions

add_device(collection, device_id, params)

Method for Add device to collection endpoint.

Add device to specified Collection.

coll: M2X.Collection struct

device_id: ID of the Device being added to Collection

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details

create!(client, params \\ %{})

Create a new resource using the given client and optional params, returning a struct with the attrs of the new resource.

delete!(collection)

Delete the remote resource.

fetch(client, id)

Method for View Collection Details endpoint.

Retrieve a view of the Collection associated with the given unique id.

client: M2X.Client struct

id: ID of the Collection to retrieve

Returns: The matching Collection

get_metadata_field(collection, name)

Method for Read Collection Metadata Field endpoint.

Get the custom metadata for the specified Collection.

coll: M2X.Collection struct

name: The metadata field to be read

Returns: The API response, see M2X API docs for details

list(client, params \\ nil)

Method for List collections endpoint.

Retrieve the list of Collections accessible by the authenticated API key that meet the search criteria.

client: M2X.Client struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: List of Device Distributions

metadata(collection)

Method for Read Collection Metadata endpoint.

Get the custom metadata for the specified Collection.

coll: M2X.Collection struct

Returns: User defined metadata associated with the collection

path(uid)

Return the API path of the Resource.

refreshed(collection)

Query the service and return a refreshed version of the same resource struct with all attrs set to their latest values.

remove_device(collection, device_id, params)

Method for Remove device from collection endpoint.

Remove device from secified Collection.

coll: M2X.Collection struct

device_id: ID of the Device being removed from Collection

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details

set_metadata_field(collection, name, value)

Method for Update Collection Metadata Field endpoint.

Update the custom metadata for the specified Collection.

coll: M2X.Collection struct

name: The metadata field to be updated

value: The value to update

Returns: The API response, see M2X API docs for details

update!(collection, params)

Update the remote resource using the given attrs.

update_metadata(collection, params)

Method for Update Collection Metadata endpoint.

Update the custom metadata for the specified Collection.

coll: M2X.Collection struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details