m2x v2.0.0 M2X.Command

Wrapper for the AT&T M2X Commands API

Functions

fetch(client, id)

Method for View Command Details endpoint.

Retrieve a view of the Command associated with the given command id.

client: M2X.Client struct

id: ID of the Command to retrieve

Returns: The matching Command

list(client, params \\ nil)

Method for List Commands endpoint.

Retrieve the list of Commands accessible by the authenticated API key.

client: M2X.Client struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: List of Commands

path(uid)

Return the API path of the Command.

send(client, params)

Method for Send Command endpoint.

Send a Command with the given name and data to the given targets.

client: M2X.Client struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The Command that was just sent