m2x v2.0.0 M2X.Device

Wrapper for the AT&T M2X Device API

Functions

catalog(client, params \\ nil)

Method for List/Search Public Devices Catalog endpoint.

This allows unauthenticated users to search Devices from other users that have been marked as public, allowing them to read public Device metadata, locations, streams list, and view each Devices’ stream metadata and its values.

client: M2X.Client struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters.

Returns: List of Devices

command(device, command_id)

Method for Device’s View of Command Details endpoint.

Get details of a received command for this Device.

device: M2X.Device struct

command_id: ID of the Command to retrieve

Returns: The API response, see M2X API docs for details

commands(device)

Method for Device’s List of Received Commands endpoint.

Retrieve the list of recent commands sent to the Device.

device: M2X.Device struct

Returns: The API response, see M2X API docs for details

create!(client, params \\ %{})

Create a new resource using the given client and optional params, returning a struct with the attrs of the new resource.

create_stream(a, b, c)

delete!(device)

Delete the remote resource.

delete_location_history(device, params)

Method for Delete Location History endpoint.

Delete location history of the specified device.

device: M2X.Device struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details

fetch(client, id)

Method for View Device Details endpoint.

Retrieve a view of the Device associated with the given unique id.

client: M2X.Client struct

id: ID of the Device to retrieve

Returns: The matching Device

get_location(device)

Method for Read Device Location endpoint.

Get location details of an existing Device. Note that this method can return an empty value (response status of 204) if the device has no location defined.

device: M2X.Device struct

Returns: Most recently logged location of the Device, see M2X API docs for details

get_metadata_field(device, name)

Method for Read Device Metadata Field endpoint.

Get the custom metadata for the specified Device.

device: M2X.Device struct

name: The metadata field to be read

Returns: The API response, see M2X API docs for details

list(client, params \\ nil)

Method for List Devices endpoint.

Retrieve the list of Devices accessible by the authenticated API key.

client: M2X.Client struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters.

Returns: List of Devices

location_history(device, params \\ %{})

Method for Read Device Location History endpoint.

Get location history details of an existing Device.

device: M2X.Device struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: Location history of the Device

metadata(device)

Method for Read Device Metadata endpoint.

Get the custom metadata for the specified Device.

device: M2X.Device struct

Returns: User defined metadata associated with the device

path(uid)

Return the API path of the Resource.

post_update(device, params)

Method for Post Device Update (Single Values to Multiple Streams) endpoint.

This method allows posting single values to multiple streams.

device: M2X.Device struct

params: The values being posted, formatted according to the API docs

Returns: The API response, see M2X API docs for details

post_updates(device, params)

Method for Post Device Update (Multiple Values to Multiple Streams) endpoint.

This method allows posting multiple values to multiple streams belonging to a device and optionally, the device location.

device: M2X.Device struct

params: The values being posted, formatted according to the API docs

Returns: The API response, see M2X API docs for details

process_command(device, command, params \\ %{})

Method for Device Marks a Command as Processed endpoint. Mark the given command as processed by this Device.

device: M2X.Device struct

id: ID of the Command being marked as processed

Returns: The API response, see M2X API docs for details

refreshed(device)

Query the service and return a refreshed version of the same resource struct with all attrs set to their latest values.

reject_command(device, command, params \\ %{})

Method for Device Marks a Command as Rejected endpoint.

Mark the given command as rejected by this Device.

device: M2X.Device struct

id: ID of the Command being marked as rejected

Returns: The API response, see M2X API docs for details

search(client, params \\ nil)

Method for Search Devices endpoint.

Retrieve the list of Devices accessible by the authenticated API key that meet the search criteria.

client: M2X.Client struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters.

Returns: List of Devices

set_metadata_field(device, name, value)

Method for Update Device Metadata Field endpoint.

Update the custom metadata field for the specified Device.

device: M2X.Device struct

name: The metadata field to be updated

value: The value to update

Returns: The API response, see M2X API docs for details

stream(device, name)

Method for View Data Stream endpoint.

Get details of a specific Stream associated with the Device.

device: M2X.Device struct

name: The name of the Stream being retrieved

Returns: The matching Stream

streams(device)

Method for List Data Streams endpoint.

Retrieve list of Streams associated with the specified Device.

device: M2X.Device struct

Returns: List of data streams associated with this device as Stream objects

update!(device, params)

Update the remote resource using the given attrs.

update_location(device, params)

Method for Update Device Location endpoint.

Update the current location of the specified device.

device: M2X.Device struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details

update_metadata(device, params)

Method for Update Device Metadata endpoint.

Update the custom metadata for the specified Device.

device: M2X.Device struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details

update_stream(device, name, params)

Method for Create/Update Data Stream endpoint.

Update a Stream associated with the Device with the given parameters. If a Stream with this name does not exist it will be created.

device: M2X.Device struct

name: Name of the stream to be updated

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The Stream being updated

values(device, params)

Method for List Values from all Data Streams of a Device endpoint.

device: M2X.Device struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details

values_export_csv(device, params \\ %{})

Method for Export Values from all Data Streams of a Device endpoint.

device: M2X.Device struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details

values_search(device, params)

Method for Search Values from all Data Streams of a Device endpoint.

device: M2X.Device struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details