m2x v2.0.0 M2X.Distribution

Wrapper for the AT&T M2X Distribution API

Functions

add_device(distribution, serial)

Method for Add Device to an existing Distribution endpoint.

Add a new Device to the Distribution, with the given unique serial string.

dist: M2X.Distribution struct

serial: The unique (account-wide) serial for the DistributionDevice being added to the Distribution

Returns: The newly created DistributionDevice

create!(client, params \\ %{})

Create a new resource using the given client and optional params, returning a struct with the attrs of the new resource.

delete!(distribution)

Delete the remote resource.

devices(distribution, params \\ nil)

Method for List Devices from an existing Distribution endpoint.

Retrieve list of Devices added to the specified Distribution.

dist: M2X.Distribution struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: List of Devices associated with this Distribution

fetch(client, id)

Method for View Distribution Details endpoint. Retrieve a view of the Distribution associated with the given unique id.

client: M2X.Client struct

id: ID of the Distribution to retrieve

Returns: The matching Distribution

get_metadata_field(distribution, name)

Method for Read Distribution Metadata Field endpoint. Get the custom metadata for the specified Distribution.

dist: M2X.Distribution struct

name: The metadata field to be read

Returns: The API response, see M2X API docs for details

list(client, params \\ nil)

Method for List Distributions endpoint.

Retrieve the list of Distributions accessible by the authenticated API key that meet the search criteria.

client: M2X.Client struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: List of Device Distributions

metadata(distribution)

Method for Read Distribution Metadata endpoint. Get the custom metadata for the specified Distribution.

dist: M2X.Distribution struct

Returns: User defined metadata associated with the distribution

path(uid)

Return the API path of the Resource.

refreshed(distribution)

Query the service and return a refreshed version of the same resource struct with all attrs set to their latest values.

set_metadata_field(distribution, name, value)

Method for Update Distribution Metadata Field endpoint. Update the custom metadata for the specified Distribution.

dist: M2X.Distribution struct

name: The metadata field to be updated

value: The value to update

Returns: The API response, see M2X API docs for details

update!(distribution, params)

Update the remote resource using the given attrs.

update_metadata(distribution, params)

Method for Update Distribution Metadata endpoint. Update the custom metadata for the specified Distribution.

dist: M2X.Distribution struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details