m2x v2.0.0 M2X.Key

Wrapper for the AT&T M2X Keys API

Functions

create!(client, params \\ %{})

Create a new resource using the given client and optional params, returning a struct with the attrs of the new resource.

delete!(key)

Delete the remote resource.

fetch(client, key)

Method for View Key Details endpoint.

Retrieve a view of the Key associated with the given unique key string.

client: M2X.Client struct

key: ID of the Key to retrieve

Returns: The matching Key

list(client, params \\ nil)

Method for List Keys endpoint.

Retrieve the list of keys associated with the specified account that meet the search criteria.

client: M2X.Client struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: List of Key

path(uid)

Return the API path of the Resource.

refreshed(key)

Query the service and return a refreshed version of the same resource struct with all attrs set to their latest values.

regenerated(key)

Method for Regenerate Key endpoint.

Regenerate an API Key token and return the regenerated Key.

Note that if you regenerate the key that you’re using for authentication then you would need to change your scripts to start using the new key token for all subsequent requests.

update!(key, params)

Update the remote resource using the given attrs.