m2x v2.0.0 M2X.Stream

Wrapper for the AT&T M2X Streams API

Functions

delete!(stream)

Delete the remote subresource.

delete_values!(stream, start, stop)

Method for Delete Data Stream Values endpoint.

Delete values in a Stream by a date range.

stream: M2X.Stream struct

start: ISO8601 timestamp for starting timerange for values to be deleted

stop: ISO8601 timestamp for ending timerange for values to be deleted

Returns: The API response, see M2X API docs for details

path(stream)

Return the API path of the Subresource.

path(under, uid)

post_values(stream, values)

Method for Post Data Stream Values endpoint.

Post a list of multiple values with timestamps to the Stream.

stream: M2X.Stream struct

values: Values to post, see M2X API docs for details

Returns: The API response, see M2X API docs for details

refreshed(stream)

Query the service and return a refreshed version of the same subresource struct with all attrs set to their latest values.

sampling(stream, params \\ nil)

Method for List Data Stream Sampling endpoint.

Sample values from the Stream matching the given optional search parameters.

stream: M2X.Stream struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details

stats(stream, params \\ nil)

Method for List Data Stream Stats endpoint.

Get statistics calculated from the values of the Stream, with optional parameters.

stream: M2X.Stream struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details

update!(stream, params)

Update the remote subresource using the given attrs.

update_value(stream, value)

Method for Update Data Stream Value endpoint.

Update the current value of the stream, with optional timestamp.

stream: M2X.Stream struct

value: The updated stream value

Returns: The API response, see M2X API docs for details

update_value(stream, value, timestamp)

values(stream, params \\ nil)

Method for List Data Stream Values endpoint.

List values from the Stream matching the given optional search parameters.

stream: M2X.Stream struct

params: Query parameters passed as keyword arguments. View M2X API Docs for listing of available parameters

Returns: The API response, see M2X API docs for details